Permit verifications via pyclassify
PyClassify now advanced enough that it runs via argparse.
Have to be careful that the matching between it and paraview and my
script remain consistent, but it is currently consistent.
Only accepts certain formats from the conversion python script for ease
of development right now.
Displays ground truth numbers for class counts when verify file is
provided.
Shows confusion when verify file is provided.
If no confusion, get a happy message and pat yourself on the back.

To this end, I swapped the min/max enum values in critPoints.cu and
fixed a bug where min/max points were always being counted as insane :/