Reduce transfer demand
Several host->device startup transfers carried "0" data as paranoia, but
CUDA guarantees this will be the case without me doing so and that saves
a significant amount of startup latency.

Still need to reduce overall memory footprint / demand, but it's a step
in the right direction.

Also includes timed context creation taken from a forum post and
commented out normal host-side malloc's for memory. We don't re-use
transfer buffers but experiments show that not-pinning the memory
doesn't produce a performance benefit yet, and we may eventually segment
the mesh such that we would re-use buffers -- best to have them pinned
for now.